projectschool-python/requirements.txt
flask
mysql-connector-python
selenium
chromedriver-binary
pytest
requests

